Allow adding events and notes without being logged in

+++ src/store/events.ts
@@ -5,9 +5,11 @@ import { computed } from 'vue';
 import { useAppStore } from './app';
 import type { Event } from '../api/api.types';
 import { format } from 'date-fns';
+import { useGeoStore } from './geo';
 
 export const useEventStore = defineStore('events', () => {
   const appStore = useAppStore();
+  const geoStore = useGeoStore();
 
   const events = ref<Event[]>([]);
 
@@ -25,19 +27,48 @@ export const useEventStore = defineStore('events', () => {
     events.value = [];
   };
 
-  const addEvent = async ({ name, date, background, cityId }: { name: string; date: string; background: string; cityId: number }) => {
+  const addEvent = async ({
+    name,
+    date,
+    background,
+    countryId,
+    cityId,
+  }: {
+    name: string;
+    date: string;
+    background: string;
+    countryId: number;
+    cityId: number;
+  }) => {
     let event: Event;
     if (appStore.isLoggedIn) {
       event = await api.addEvent({ name, date, background, cityId }, { authToken: appStore.user?.token });
     } else {
-      event = {} as any;
+      const country = geoStore.getCountryById(countryId);
+      if (!country) throw new Error(`Unable to find country: ${countryId}`);
+      const city = geoStore.getCityById(cityId);
+      if (!city) throw new Error(`Unable to find city: ${cityId}`);
+      const randomIds = new Uint32Array(1);
+      crypto.getRandomValues(randomIds);
+
+      event = {
+        id: randomIds[0],
+        name,
+        eventDate: new Date(date),
+        order: events.value.length + 1,
+        background,
+        country,
+        city,
+      };
     }
 
     events.value.push(event);
   };
 
   const removeEvent = async (eventId: number) => {
-    await api.deleteEvent({ eventId }, { authToken: appStore.user?.token });
+    if (appStore.isLoggedIn) {
+      await api.deleteEvent({ eventId }, { authToken: appStore.user?.token });
+    }
     const index = events.value.findIndex((event) => event.id === eventId);
     events.value.splice(index, 1);
   };
@@ -47,28 +78,51 @@ export const useEventStore = defineStore('events', () => {
     name,
     date,
     background,
+    countryId,
     cityId,
   }: {
     eventId: number;
     name: string;
     date: Date;
     background: string;
+    countryId: number;
     cityId: number;
   }) => {
-    const eventDate = format(date, 'yyyy-MM-dd HH:mm:00');
-    const event = await api.updateEvent({ eventId, name, date: eventDate, background, cityId }, { authToken: appStore.user?.token });
+    let event: Event;
+    if (appStore.isLoggedIn) {
+      const eventDate = format(date, 'yyyy-MM-dd HH:mm:00');
+      event = await api.updateEvent({ eventId, name, date: eventDate, background, cityId }, { authToken: appStore.user?.token });
+    } else {
+      const country = geoStore.getCountryById(countryId);
+      if (!country) throw new Error(`Unable to find country: ${countryId}`);
+      const city = geoStore.getCityById(cityId);
+      if (!city) throw new Error(`Unable to find city: ${cityId}`);
+
+      event = {
+        id: eventId,
+        name,
+        eventDate: new Date(date),
+        order: events.value.length + 1,
+        background,
+        country,
+        city,
+      };
+    }
     const eventIndex = events.value.findIndex((e) => e.id === event.id);
     events.value.splice(eventIndex, 1, event);
   };
 
   const reorderEvents = async (newlyOrderedEvents: Event[]) => {
-    isReorderingEvents.value = true;
+    if (appStore.isLoggedIn) isReorderingEvents.value = true;
 
     const originalEventOrder: Event[] = events.value;
     events.value = newlyOrderedEvents.map((e, index) => {
       e.order = index + 1;
       return e;
     });
+
+    if (!appStore.isLoggedIn) return;
+
     const newOrderIds = newlyOrderedEvents.map((event) => event.id);
     try {
       const _events = await api.reorderEvents({ eventIds: newOrderIds }, { authToken: appStore.user?.token });
